deleted old Challenges class and remade it as ChallengesService
